+++ aleo/artifacts/js/js2leo/token_service.ts
@@ -8,7 +8,7 @@ import {
 import * as js2leo from "./common";
 export function getTSForeignContractLeo(tSForeignContract: TSForeignContract): TSForeignContractLeo {
   const result: TSForeignContractLeo = {
-    chain_id: js2leo.u32(tSForeignContract.chain_id),
+    chain_id: js2leo.u128(tSForeignContract.chain_id),
     contract_address: js2leo.array(tSForeignContract.contract_address, js2leo.u8),
   }
   return result;
@@ -16,7 +16,7 @@ export function getTSForeignContractLeo(tSForeignContract: TSForeignContract): T
 
 export function getTokenOriginLeo(tokenOrigin: TokenOrigin): TokenOriginLeo {
   const result: TokenOriginLeo = {
-    chain_id: js2leo.u32(tokenOrigin.chain_id),
+    chain_id: js2leo.u128(tokenOrigin.chain_id),
     token_service_address: js2leo.array(tokenOrigin.token_service_address, js2leo.u8),
     token_address: js2leo.array(tokenOrigin.token_address, js2leo.u8),
   }
